chama_system/requirements.txt
Django==5.0.14
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
psycopg2-binary==2.9.11
python-decouple==3.8
beautifulsoup4==4.14.3
python-docx==1.2.0
lxml==6.0.2
pillow==12.1.0
reportlab==4.4.9
soupsieve==2.8.3
